# app/goods/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Goods API
    version: 0.0.1
paths:
    /v1/goods:
        get:
            tags:
                - Goods
            operationId: Goods_ListGoods
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListGoodsReply'
    /v1/goods/{id}:
        get:
            tags:
                - Goods
            description: |-
                rpc CreateGoods (CreateGoodsRequest) returns (CreateGoodsReply){
                 	option (google.api.http) = {
                 		post: "/v1/goods"
                 		body: "*"
                 	};
                 };
                 rpc UpdateGoods (UpdateGoodsRequest) returns (UpdateGoodsReply){
                 	option (google.api.http) = {
                 		put: "/v1/goods"
                 		body: "*"
                 	};
                 };
                 rpc DeleteGoods (DeleteGoodsRequest) returns (DeleteGoodsReply){
                 	option (google.api.http) = {
                 		delete: "/v1/goods/{id}"
                 	};
                 };
            operationId: Goods_GetGoods
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetGoodsReply'
components:
    schemas:
        GetGoodsReply:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                name:
                    type: string
                price:
                    type: number
                    format: float
        Item:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                name:
                    type: string
                price:
                    type: number
                    format: float
        ListGoodsReply:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/Item'
tags:
    - name: Goods